azure-intelligent-search-application/azure-intelligent-search-application/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1

azure-core==1.30.1
azure-search-documents==11.6.0
azure-storage-blob==12.21.0

scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.5

python-docx==1.1.0
PyPDF2==3.0.1
requests==2.32.3
matplotlib==3.9.2
umap-learn==0.5.6
